Add markers css class (#2)
## Background

We want to be able to hightlight lines in the code block, but this is
not supported out of the box and requires custom styling.

## Solution

Move the CodeBlockResolver to its own folder, and create a css file that
holds the styling for the `refractor-marker` css class. This class is
used to override the syntax highlighting. As this file is relevant to
show code examples of, it also makes sense to register `css` as a
language from refractor in the process.