Add TypeGen types from Sanity and refactored types to support this (#4)
## Background

We want to use proper typing from Sanity to sync the data structure in
the data source and frontend properly!

## Solution

Imported the TypeGen types from Sanity, and refactored the names and
validation of fields to match this change. I think this will be a way
more consistent and powerful approach going forward!