Fix mobile presentation (#7)
## Background

For some reason all the styling and elements looks horrible on mobile.
Text overflows, the background only covers half and everything is just
crazy.

## Solution

Fix a hardcoded width that should have been a maximum. Also, removed the
default margin on the body due to creating a weird border due to
different background colors.