Create script for creating components with specific styling type (#8)
## Background

Creating new components can be kinda tedious, as there are manual
efforts needed to setup basic boilerplate for tsx, styling and a barrel
file.

## Solution

Create an awesome script for creating components, that creates the
standard files needed for my basic components. To run you only need to
run `./createComponent.ts <name> <styling>` in the `scripts` directory
to do it. This should make it extendable to use other styling technices
than just css that is the only supported one now, and is a part of the
path to the correct component directory.

It also provides the basic code for both the barrel and exported
component for const. I have chosen a bare-bones approach, so you'll have
to add correct implementation instead of deleting unnecessary
boilterplate.

It has some great user-friendy functionality too:
* Does not allow the creation of components that already exists
* Gives you explanation of usage if you provide missing arguments
* Renames the given component name to fit repo's code style
* Tells you what styling options are supported
* Gives a confirmation on successful creation, and gives error report on
unsuccessful