Rich image list (#9)
## Background

We want to be able to show images, with metadata for accessibility and
attribution. This should also support the custom image type created for
articles in Sanity.

## Solution

Created RichImage component, that creates an Image component. I have
tried to achieve semantic cohesion by using the `<figure>` and
`<figcaption>` tags when there is a caption or credit attached to the
image.

The resolver for the Sanity type uses a new package `@sanity/image-url`
to build custom URLs to fetch the image from Sanity. This adds support
for features like hotspot and cropping. By combining this with the
srcSet and sizes properties for the `<img>` tag, we also control what
quality of the image we want to load based on screen size and
retina-displays. This is an optimization that hopefully 🤞 will save us
some bandwidth, even though this isn't an issue so far.

To make this look nice I also added the support for multiple text
styles. Currently the new ones are secondary, muted and highlight. All
will be very useful, but the colors will probably need some tweaking
when the design system matures.