Use refactored CodeBlock type from Sanity (#10)
We want to be able to show metadata for code blocks, which have been
implemented in Sanity to present a filename if specified.

Ran TypeGen from Sanity again with the new type, and refactored the
Resolver to use this instead. Now we show a filename in the bottom right
corner of the code block if it has been defined.