Add copy button to code blocks (#11)
## Background

People will most likely want to copy the code you're presenting. This
should be as user friendly as a click of a button.

## Solution

Created a hook for interacting with the clipboard: `useClipboard`. I
have added a button using a copy icon from the `react-icons` library
I've added. This allows the user to get the code directly to their
clipboard if the browser supports it. No fallback method added.

Also fixed some styling in different components.